Refactored Mindhunters (models, web app)
