+++ app.py
@@ -1,98 +1,15 @@
+import os
+from functools import wraps
 from flask import Flask, flash, jsonify, redirect, render_template, request, session
 from flask_session import Session
-from functools import wraps
-import os
-from cs50 import SQL
 from tempfile import mkdtemp
+from cs50 import SQL
+from tensorflow.keras.models import load_model
 from werkzeug.utils import secure_filename
 from werkzeug.exceptions import default_exceptions, HTTPException, InternalServerError
 from werkzeug.security import check_password_hash, generate_password_hash
 
-#######
-import pickle
-import pandas as pd
-import numpy as np
-import string
-import nltk
-from nltk.stem import WordNetLemmatizer
-from nltk.corpus import stopwords
-from nltk.corpus import wordnet
-from nltk import pos_tag
-from nltk.tokenize import word_tokenize
-from tensorflow.keras.models import load_model
-
-from keras.models import Model
-from keras.layers import Dense, Input, Dropout, LSTM, Activation
-from keras.layers.embeddings import Embedding
-from keras.preprocessing import sequence
-from keras.initializers import glorot_uniform
-
-lemmatizer = WordNetLemmatizer()
-stop_words = set(stopwords.words('english'))
-stop_words.update(list(string.punctuation))
-
-def get_simple_pos(tag) :
-    if tag.startswith('J') :
-        return wordnet.ADJ
-    elif tag.startswith('V') :
-        return wordnet.VERB
-    elif tag.startswith('N') :
-        return wordnet.NOUN
-    elif tag.startswith('R') :
-        return wordnet.ADV
-    else:
-        return wordnet.NOUN
-
-max_len = 30
-def clean_text(review) :
-    global max_len 
-    words = word_tokenize(review)
-    output_words = []
-    for word in words :
-        if word.lower() not in stop_words :
-            pos = pos_tag([word])
-            clean_word = lemmatizer.lemmatize(word,pos = get_simple_pos(pos[0][1]))
-            output_words.append(clean_word.lower())
-    max_len = max(max_len, len(output_words))
-    return " ".join(output_words)
-
-def read_glove_vecs(glove_file):
-    with open(glove_file, 'r', encoding="utf8") as file:
-        word_to_vec_map = {}
-        word_to_index = {}
-        index_to_word = {}
-        index = 0
-        for line in file:
-            line = line.strip().split()
-            curr_word = line[0]
-            word_to_index[curr_word] = index
-            index_to_word[index] = curr_word
-            word_to_vec_map[curr_word] = np.array(line[1:], dtype=np.float64)
-            index += 1
-    return word_to_index, index_to_word, word_to_vec_map
-
-def sentences_to_indices(X, word_to_index, max_len):
-    m = len(X)
-    X_indices = np.zeros((m, max_len))
-    for i in range(m):
-        sentence_words = [w.lower() for w in X[i].split()]
-        j = 0
-        for word in sentence_words:
-            if word in word_to_index:
-                X_indices[i, j] = word_to_index[word]
-            j += 1
-    return X_indices
-
-#word_to_index, index_to_word, word_to_vec_map = read_glove_vecs('glove.6B.50d.txt')
-filename = 'word_to_index.pkl'
-word_to_index =  pickle.load(open(filename, 'rb')) 
-print(len(word_to_index))
-
-model = load_model('model.h5')
-#print(model.summary())
-
-#######
-
+from src import mindhunters
 
 # Configure application
 app = Flask(__name__)
@@ -104,7 +21,10 @@ app.config["SESSION_PERMANENT"] = False
 app.config["SESSION_TYPE"] = "filesystem"
 Session(app)
 # Configure to use SQLite database
-db = SQL("sqlite:///main.db")
+db = SQL("sqlite:///src/main.db")
+# Load the ML model and initiate Mindhunters
+model = load_model('src/model.h5')
+word_to_index, max_len = mindhunters.init()
 
 def login_required(f):
     @wraps(f)
@@ -189,9 +109,8 @@ def index():
         post_text = request.form.get("post")
         if not post_text:
             return redirect("/")
-        text = clean_text(post_text)
-        text = [text]
-        text = sentences_to_indices(text,word_to_index,max_len)
+        text = [mindhunters.clean_text(post_text)]
+        text = mindhunters.sentences_to_indices(text, word_to_index, max_len)
         ans = model.predict(text)[0][0]
         db.execute("INSERT INTO :tablename ('text', 'nature') VALUES (:post_text, :score)", tablename=userInfo['username'], post_text=post_text, score=str(ans))
         if (ans < 0.4):
